Updated XML and Searching
Updated XML to fix the appearance of the button, text no longer spills to a second line and the button does not touch the edge of the screen. Also changed the XML to accommodate a popup at the bottom of the screen. Changed PlayedSearch code such that when a user does not enter any text, the search does not run and a popup message appears to tell the user to enter text. Also added a popup so that if there are no matching users to the player's search, a pop up displays a message indicating so.