Change image generation method to use Hash
Change the image generation method to use the Hash value of the code. Specifically, each pixel now sets its ARGB value based on the hash of the code by taking 4-character chunks of the hash, adding the X+Y+Offset coordinates of the pixel, % 256 that whole value, and adding that as the pixels A R G B values.