Create Container for Hash Objects
Create a custom class for containing a code hash, so that we can have a constructor in Code() that takes this class (which just contains a string), and builds a Code object from it. Essentially build a Code object using a pre-made hash instead of taking a string and creating a new hash from it.

The new class is required because the Data and the Hash are both strings, so constructors would get confused.